requirements.txt
aiofiles==23.2.1
alembic==1.12.0
anyio==3.3.4
asgiref==3.4.1
attrs==23.1.0
bcrypt==3.2.0
boto3==1.20.13
botocore==1.23.13
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
click==8.0.3
colorama==0.4.6
decorator==4.4.2
dnspython==2.4.2
ecdsa==0.17.0
email-validator==2.0.0.post2
exceptiongroup==1.2.1
fastapi==0.70.0
fastapi-cli==0.0.3
greenlet==1.1.2
httptools==0.2.0
httpx==0.27.0
idna==3.3
imageio==2.34.1
imageio-ffmpeg==0.4.9
iniconfig==2.0.0
Jinja2==3.1.4
jmespath==0.10.0
Mako==1.2.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
moviepy==1.0.3
numpy==1.21.4
opencv-python==4.9.0.80
orjson==3.10.3
packaging==23.2
pandas==1.3.4
passlib==1.7.4
pillow==10.3.0
pluggy==1.3.0
proglog==0.1.10
psycopg2-binary==2.9.3
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.8.2
Pygments==2.18.0
pytest==7.1.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.9
pytz==2021.3
PyYAML==6.0.1
requests==2.26.0
rich==13.7.1
rsa==4.7.2
s3transfer==0.5.0
shellingham==1.5.4
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.27
starlette==0.16.0
tomli==2.0.1
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.0.0
ujson==5.9.0
urllib3==1.26.7
uvicorn==0.15.0
watchgod==0.8.2
websockets==12.0
